refactor: move TopLevelClientLogic to a shared location and update imports
